Refactor task schemas and services to align with new module structure
