+++ src/api/v1/endpoints/benchmark_task.py
@@ -2,12 +2,12 @@ from fastapi import APIRouter, Depends, Query
 from sqlalchemy.orm import Session
 
 from src.api.deps import api_key_header
-from src.api.v1.schemas.tasks.benchmark_task import (
+from src.api.v1.schemas.tasks.benchmark.benchmark_task import (
     BenchmarkCreate,
     BenchmarkCreateResponse,
     BenchmarkResponse,
 )
-from src.api.v1.schemas.tasks.device import SupportedDevicesForBenchmarkResponse
+from src.api.v1.schemas.tasks.common.device import SupportedDevicesForBenchmarkResponse
 from src.core.db.session import get_db
 from src.services.benchmark_task import benchmark_task_service
