+++ src/api/v1/schemas/tasks/common/device.py
