+++ src/api/v1/schemas/tasks/compression/compression_task.py
