+++ src/api/v1/schemas/tasks/training/augmentation.py
