+++ src/api/v1/schemas/tasks/training/environment.py
